audio-hal : Update data type to avoid data corruption.
Data type is modified to size_t as it causes data corruption
in 64 bit builds.

Suggested-by : Manoj Kumar N D

+++ hal/audio_extn/utils.c
@@ -1046,7 +1046,7 @@ static int set_stream_app_type_mixer_ctrl(struct audio_device *adev,
 
     char mixer_ctl_name[MAX_LENGTH_MIXER_CONTROL_IN_INT] = {0};
     struct mixer_ctl *ctl;
-    int app_type_cfg[MAX_LENGTH_MIXER_CONTROL_IN_INT] = {0}, len = 0, rc = 0;
+    size_t app_type_cfg[MAX_LENGTH_MIXER_CONTROL_IN_INT] = {0}, len = 0, rc = 0;
     int snd_device_be_idx = -1;
 
     if (stream_type == PCM_PLAYBACK) {
